Finish cleaning up the permissions.

+++ src/AppBundle/Controller/PersonController.php
@@ -122,6 +122,10 @@ class PersonController extends Controller {
      * @param Request $request
      */
     public function newAction(Request $request) {
+        if( ! $this->isGranted('ROLE_BLOG_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $person = new Person();
         $form = $this->createForm('AppBundle\Form\PersonType', $person);
         $form->handleRequest($request);
@@ -166,6 +170,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function editAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_BLOG_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $editForm = $this->createForm('AppBundle\Form\PersonType', $person);
         $editForm->handleRequest($request);
 
@@ -191,6 +199,10 @@ class PersonController extends Controller {
      * @param Person $person
      */
     public function deleteAction(Request $request, Person $person) {
+        if( ! $this->isGranted('ROLE_BLOG_ADMIN')) {
+            $this->addFlash('danger', 'You must login to access this page.');
+            return $this->redirect($this->generateUrl('fos_user_security_login'));
+        }
         $em = $this->getDoctrine()->getManager();
         $em->remove($person);
         $em->flush();
